[cmd] Start deprecating old install logic
